Ported to new API, added common arguments

+++ src/java/org/commoncrawl/examples/ExampleMetadataDomainPageCount.java
@@ -1,6 +1,5 @@
 package org.commoncrawl.examples;
 
-// Java classes
 import java.io.IOException;
 import java.net.URI;
 
@@ -8,20 +7,16 @@ import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.conf.Configured;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
-import org.apache.hadoop.fs.PathFilter;
 import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
-import org.apache.hadoop.mapred.FileInputFormat;
-import org.apache.hadoop.mapred.FileOutputFormat;
-import org.apache.hadoop.mapred.JobClient;
-import org.apache.hadoop.mapred.JobConf;
-import org.apache.hadoop.mapred.MapReduceBase;
-import org.apache.hadoop.mapred.Mapper;
-import org.apache.hadoop.mapred.OutputCollector;
-import org.apache.hadoop.mapred.Reporter;
-import org.apache.hadoop.mapred.SequenceFileInputFormat;
-import org.apache.hadoop.mapred.TextOutputFormat;
-import org.apache.hadoop.mapred.lib.LongSumReducer;
+import org.apache.hadoop.mapreduce.Job;
+import org.apache.hadoop.mapreduce.Mapper;
+import org.apache.hadoop.mapreduce.lib.input.FileInputFormat;
+import org.apache.hadoop.mapreduce.lib.input.SequenceFileInputFormat;
+import org.apache.hadoop.mapreduce.lib.output.FileOutputFormat;
+import org.apache.hadoop.mapreduce.lib.output.TextOutputFormat;
+import org.apache.hadoop.mapreduce.lib.reduce.LongSumReducer;
+import org.apache.hadoop.util.GenericOptionsParser;
 import org.apache.hadoop.util.Tool;
 import org.apache.hadoop.util.ToolRunner;
 import org.apache.log4j.Logger;
@@ -29,10 +24,6 @@ import org.apache.log4j.Logger;
 import com.google.common.net.InternetDomainName;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
-// Apache Project classes
-// Hadoop classes
-// Google Gson classes
-// Google Guava classes
 
 /**
  * An example showing how to use the Common Crawl 'metadata' files to quickly
@@ -40,68 +31,88 @@ import com.google.gson.JsonParser;
  * 
  * @author Chris Stephens <chris@commoncrawl.org>
  */
-public class ExampleMetadataDomainPageCount
-    extends    Configured
-    implements Tool {
+public class ExampleMetadataDomainPageCount extends Configured implements Tool {
 
   private static final Logger LOG = Logger.getLogger(ExampleMetadataDomainPageCount.class);
+  private static final String ARGNAME_INPATH = "-in";
+  private static final String ARGNAME_OUTPATH = "-out";
+  private static final String ARGNAME_CONF = "-conf";
+  private static final String ARGNAME_OVERWRITE = "-overwrite";
+  private static final String ARGNAME_MAXFILES = "-maxfiles";
+  private static final String ARGNAME_NUMREDUCE = "-numreducers";
+  private static final String FILTER_PREFIX = "metadata-";
+  
+  protected static enum MAPPERCOUNTER {
+    INVALID_URIS,
+    INVALID_DOMAIN,
+    MISSING_HTTPCODE,
+    HTTP_SUCCESS,
+    HTTP_FAIL,
+    EXCEPTIONS
+  }
 
   /**
    * Mapping class that produces the normalized domain name and a count of '1'
    * for every successfully retrieved URL in the Common Crawl corpus.
    */ 
   public static class ExampleMetadataDomainPageCountMapper
-      extends    MapReduceBase
-      implements Mapper<Text, Text, Text, LongWritable> {
+      extends Mapper<Text, Text, Text, LongWritable> {
+    
+    private final Text outKey = new Text();
+    private static final LongWritable outVal = new LongWritable(1);
+    
+    String url;
+    String json;
+    URI uri;
+    String host;
+    InternetDomainName domainName;
+    String domain;
+    
+    JsonParser jsonParser = new JsonParser();
+    JsonObject jsonObj;
 
-    // create a counter group for Mapper-specific statistics
-    private final String _counterGroup = "Custom Mapper Counters";
-
-    // implement the main "map" function
-    public void map(Text key, Text value, OutputCollector<Text, LongWritable> output, Reporter reporter)
+    @Override
+    public void map(Text key, Text value, Context context)
         throws IOException {
 
       // key & value are "Text" right now ...
-      String url   = key.toString();
-      String json  = value.toString();
+      url   = key.toString();
+      json  = value.toString();
 
       try {
 
         // Get the base domain name
-        URI uri = new URI(url);
-        String host = uri.getHost();
+        uri = new URI(url);
+        host = uri.getHost();
 
         if (host == null) {
-          reporter.incrCounter(this._counterGroup, "Invalid URI", 1);
+          context.getCounter(MAPPERCOUNTER.INVALID_URIS).increment(1);
           return;
         }
 
-        InternetDomainName domainObj = InternetDomainName.from(host);
-
-        String domain = domainObj.topPrivateDomain().name();
+        domainName = InternetDomainName.from(host);
+        domain = domainName.topPrivateDomain().name();
 
         if (domain == null) {
-          reporter.incrCounter(this._counterGroup, "Invalid Domain", 1);
+          context.getCounter(MAPPERCOUNTER.INVALID_DOMAIN).increment(1);
           return;
         }
 
         // See if the page has a successful HTTP code
-        JsonParser jsonParser = new JsonParser();
-        JsonObject jsonObj    = jsonParser.parse(json).getAsJsonObject();
+        jsonObj = jsonParser.parse(json).getAsJsonObject();
 
         if (jsonObj.has("http_result") == false) {
-          reporter.incrCounter(this._counterGroup, "HTTP Code Missing", 1);
+          context.getCounter(MAPPERCOUNTER.MISSING_HTTPCODE).increment(1);
           return;
         }
 
         if (jsonObj.get("http_result").getAsInt() == 200) {
-          reporter.incrCounter(this._counterGroup, "HTTP Success", 1);
-
-          // only output counts for pages that were successfully retrieved
-          output.collect(new Text(domain), new LongWritable(1));
+          context.getCounter(MAPPERCOUNTER.HTTP_SUCCESS).increment(1);
+          outKey.set(domain);
+          context.write(outKey, outVal);
         }
         else {
-          reporter.incrCounter(this._counterGroup, "HTTP Not Success", 1);
+          context.getCounter(MAPPERCOUNTER.HTTP_FAIL).increment(1);
         }
       }
       catch (IOException ex) {
@@ -109,36 +120,21 @@ public class ExampleMetadataDomainPageCount
       }
       catch (Exception ex) {
         LOG.error("Caught Exception", ex); 
-        reporter.incrCounter(this._counterGroup, "Exceptions", 1);
+        context.getCounter(MAPPERCOUNTER.EXCEPTIONS).increment(1);
       }
     }
   }
-
-
-  /**
-   * Hadoop FileSystem PathFilter for ARC files, allowing users to limit the
-   * number of files processed.
-   *
-   * @author Chris Stephens <chris@commoncrawl.org>
-   */
-  public static class SampleFilter
-      implements PathFilter {
-
-    private static int count =         0;
-    private static int max   = 999999999;
-
-    public boolean accept(Path path) {
-
-      if (!path.getName().startsWith("metadata-"))
-        return false;
-
-      SampleFilter.count++;
-
-      if (SampleFilter.count > SampleFilter.max)
-        return false;
-
-      return true;
-    }
+  
+  public void usage() {
+    System.out.println("\n  org.commoncrawl.examples.ExampleMetadataDomainPageCount \n" +
+                         "                           " + ARGNAME_INPATH +" <inputpath>\n" +
+                         "                           " + ARGNAME_OUTPATH + " <outputpath>\n" +
+                         "                         [ " + ARGNAME_OVERWRITE + " ]\n" +
+                         "                         [ " + ARGNAME_NUMREDUCE + " <number_of_reducers> ]\n" +
+                         "                         [ " + ARGNAME_CONF + " <conffile> ]\n" +
+                         "                         [ " + ARGNAME_MAXFILES + " <maxfiles> ]");
+    System.out.println("");
+    GenericOptionsParser.printGenericCommandUsage(System.out);
   }
 
   /**
@@ -149,26 +145,43 @@ public class ExampleMetadataDomainPageCount
    * @return      0 if the Hadoop job completes successfully, 1 if not. 
    */
   @Override
-  public int run(String[] args)
-      throws Exception {
+  public int run(String[] args) throws Exception {
 
+    String inputPath = null;
     String outputPath = null;
     String configFile = null;
+    boolean overwrite = false;
+    int numReducers = 1;
 
-    // Read the command line arguments.
-    if (args.length <  1)
-      throw new IllegalArgumentException("Example JAR must be passed an output path.");
-
-    outputPath = args[0];
-
-    if (args.length >= 2)
-      configFile = args[1];
-
-    // For this example, only look at a single metadata file.
-    String inputPath = "s3n://aws-publicdatasets/common-crawl/parse-output/segment/1341690166822/metadata-01849";
- 
-    // Switch to this if you'd like to look at all metadata files.  May take many minutes just to read the file listing.
-    // String inputPath = "s3n://aws-publicdatasets/common-crawl/parse-output/segment/*/metadata-*";
+    // Read the command line arguments. We're not using GenericOptionsParser
+    // to prevent having to include commons.cli as a dependency.
+    for (int i = 0; i < args.length; i++) {
+      try {
+        if (args[i].equals(ARGNAME_INPATH)) {
+          inputPath = args[++i];
+        } else if (args[i].equals(ARGNAME_OUTPATH)) {
+          outputPath = args[++i];
+        } else if (args[i].equals(ARGNAME_CONF)) {
+          configFile = args[++i];
+        } else if (args[i].equals(ARGNAME_MAXFILES)) {
+          SampleFilter.setMax(Long.parseLong(args[++i]));
+        } else if (args[i].equals(ARGNAME_OVERWRITE)) {
+          overwrite = true;
+        } else if (args[i].equals(ARGNAME_NUMREDUCE)) {
+          numReducers = Integer.parseInt(args[++i]);
+        } else {
+          LOG.warn("Unsupported argument: " + args[i]);
+        }
+      } catch (ArrayIndexOutOfBoundsException e) {
+        usage();
+        throw new IllegalArgumentException();
+      }
+    }
+    
+    if (inputPath == null || outputPath == null) {
+      usage();
+      throw new IllegalArgumentException();
+    }
 
     // Read in any additional config parameters.
     if (configFile != null) {
@@ -176,25 +189,26 @@ public class ExampleMetadataDomainPageCount
       this.getConf().addResource(configFile);
     }
 
-    // Creates a new job configuration for this Hadoop job.
-    JobConf job = new JobConf(this.getConf());
-
+    // Create the Hadoop job.
+    Configuration conf = getConf();
+    Job job = new Job(conf);
     job.setJarByClass(ExampleMetadataDomainPageCount.class);
+    job.setNumReduceTasks(numReducers);
 
     // Scan the provided input path for ARC files.
     LOG.info("setting input path to '"+ inputPath + "'");
+    SampleFilter.setPrefix(FILTER_PREFIX);
     FileInputFormat.addInputPath(job, new Path(inputPath));
+    FileInputFormat.setInputPathFilter(job, SampleFilter.class);
 
-    // Optionally, you can add in a custom input path filter
-    // FileInputFormat.setInputPathFilter(job, SampleFilter.class);
-
-    // Delete the output path directory if it already exists.
-    LOG.info("clearing the output path at '" + outputPath + "'");
-
-    FileSystem fs = FileSystem.get(new URI(outputPath), job);
-
-    if (fs.exists(new Path(outputPath)))
-      fs.delete(new Path(outputPath), true);
+    // Delete the output path directory if it already exists and user wants to overwrite it.
+    if (overwrite) {
+      LOG.info("clearing the output path at '" + outputPath + "'");
+      FileSystem fs = FileSystem.get(new URI(outputPath), conf);
+      if (fs.exists(new Path(outputPath))) {
+        fs.delete(new Path(outputPath), true);
+      }
+    }
 
     // Set the path where final output 'part' files will be saved.
     LOG.info("setting output path to '" + outputPath + "'");
@@ -202,10 +216,10 @@ public class ExampleMetadataDomainPageCount
     FileOutputFormat.setCompressOutput(job, false);
 
     // Set which InputFormat class to use.
-    job.setInputFormat(SequenceFileInputFormat.class);
+    job.setInputFormatClass(SequenceFileInputFormat.class);
 
     // Set which OutputFormat class to use.
-    job.setOutputFormat(TextOutputFormat.class);
+    job.setOutputFormatClass(TextOutputFormat.class);
 
     // Set the output data types.
     job.setOutputKeyClass(Text.class);
@@ -215,10 +229,11 @@ public class ExampleMetadataDomainPageCount
     job.setMapperClass(ExampleMetadataDomainPageCount.ExampleMetadataDomainPageCountMapper.class);
     job.setReducerClass(LongSumReducer.class);
 
-    if (JobClient.runJob(job).isSuccessful())
+    if (job.waitForCompletion(true)) {
       return 0;
-    else
+    } else {
       return 1;
+    }
   }
 
   /**
